Rename unit Tests to UnitTest (#23)
