+++ .github/workflows/ci-build.yml
@@ -63,7 +63,7 @@ jobs:
           action_fail_on_inconclusive: true
           junit_files: |
             **/build/test-results/*/TEST-*.xml
-            !**/build/test-results/*/TEST-net.twisterrob.astro.bazi.SolarCalculatorTest.xml
+            !**/build/test-results/*/TEST-net.twisterrob.astro.bazi.SolarCalculatorUnitTest.xml
             !**/build/test-results/*/TEST-*UiTest.xml
 
       - name: "Publish 'UI Results' check suite."
@@ -88,7 +88,7 @@ jobs:
           test_changes_limit: 0
           action_fail_on_inconclusive: true
           junit_files: |
-            **/build/test-results/*/TEST-net.twisterrob.astro.bazi.SolarCalculatorTest.xml
+            **/build/test-results/*/TEST-net.twisterrob.astro.bazi.SolarCalculatorUnitTest.xml
 
       - name: "Upload 'Detekt Results' artifact."
         if: ${{ success() || failure() }}
