+++ modules/domain/bazi-calculator/base/src/test/kotlin/net/twisterrob/astro/bazi/lookup/EarthlyBranch_atHourKtUnitTest.kt
@@ -7,7 +7,7 @@ import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
 import org.junit.jupiter.params.provider.ValueSource
 
-class EarthlyBranch_atHourKtTest {
+class EarthlyBranch_atHourKtUnitTest {
 
 	@CsvSource(
 		"0, Zi",
