+++ modules/domain/bazi-calculator/wikipedia/src/test/kotlin/net/twisterrob/astro/bazi/SexagenaryWikiCalculatorUnitTest.kt
@@ -3,7 +3,7 @@ package net.twisterrob.astro.bazi
 import org.junit.jupiter.api.Disabled
 
 @Disabled("Partial / inconsistent implementation, months are very imprecisely approximated.")
-class SexagenaryWikiCalculatorTest : BaZiCalculatorTest() {
+class SexagenaryWikiCalculatorUnitTest : BaZiCalculatorBaseUnitTest() {
 
 	override val subject = SexagenaryWikiCalculator()
 }
