+++ modules/domain/bazi-calculator/wikipedia/src/test/kotlin/net/twisterrob/astro/bazi/lookup/EarthlyBranch_atSolarDayKtUnitTest.kt
@@ -6,7 +6,7 @@ import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
 import java.time.LocalDate
 
-class EarthlyBranch_atSolarDayKtTest {
+class EarthlyBranch_atSolarDayKtUnitTest {
 
 	@CsvSource(
 		// Edge of Gregorian year.
